Normalize client disconnect errors during patch
As this is expected behaviour during some circumstances we should
just ignore these errors. This was previously done for some ASP.NET pipelines
but not for all. The code has been updated to always silently return 200 OK if a client
disconnects. This will of course not reach the client as it has disconnected.
The important thing is to not flood the system with error logs.