Add support for checksum-trailer
No new interface is required for the store.
If the checksum-trailer is not valid (or the client disconnected)
tus dotnet will provide a fallback checksum that always fails.
Store authors can use tusdotnet.Helpers.ChecksumTrailerHelper.IsFallback
to determine if the provided algorithm and checksum is a fallback.
IsFallback will always return false for pipelines where trailing request
headers are not supported.