Lock on HEAD requests
This increases stability in cases where data is in transfer but has not yet reached the server when the client disconnects and reconnects. In earlier version this would cause the wrong offset to be retrieved. The client would continue from that offset causing an error when trying to patch the data once the lock had been released (offset mismatch).

Response code for locked operations have been changed from 409 Conflict to 423 Locked to better conform to how clients and other servers are implemented.