Dev/205 new client disconnect guard (#207)
* Add new client disconnect guard with support for timeouts

Ref: #205

* Add tests

Ref #205

* Remove support for timeouts in OWIN

+ Code cleanup and xmldoc on ClientReadTimeout

* Cleanup code and add xml docs

* Fix broken tests

* Optimize timeout handling for .NET standard

* Add App_Data/tusfiles to gitignore (properly this time)

* Rebase on master

* Lock on HEAD requests

This increases stability in cases where data is in transfer but has not yet reached the server when the client disconnects and reconnects. In earlier version this would cause the wrong offset to be retrieved. The client would continue from that offset causing an error when trying to patch the data once the lock had been released (offset mismatch).

Response code for locked operations have been changed from 409 Conflict to 423 Locked to better conform to how clients and other servers are implemented.

* Use older version of the test host on old .NET Core versions to make it work in VS

* Cleanup