Use NotApplicable intention if client uses concatenation but concatenation is not supported
This is an edge case where the client sends a request to use
concatenation without first checking if the server supports it.
Previous implementation did a fallback for creation meaning that
the partial files were created (as normal files) but the final file failed
because tusdotnet could not merge the partial files.

The new implementation returns NotApplicable causing a 404 Not Found
to be returned to the client.

This matches the behaviour with if creation is not supported.

Ref #238